show the search warning in all three places
